+++ Assets/VRM10/Editor/Components/SpringBone/VRM10SelectorWindow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a38b6d598e4f6c9458cd0a819302b438
+guid: 96fdd3a9c7c5ca44b977ef564a2cf15b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
